updated link elements to use same wrapper component
