+++ src/components/Elements/LinkElements/Link.tsx
@@ -3,7 +3,8 @@ import {
   Link as ChakraLink,
   LinkProps as ChakraLinkProps,
 } from '@chakra-ui/react'
-import NextLink, { LinkProps } from 'next/link'
+import { LinkProps } from 'next/link'
+import { LinkWrapper } from './LinkWrapper'
 
 type ChakraLinkAndNextProps = ChakraLinkProps & LinkProps
 
@@ -13,16 +14,10 @@ const Link = ({
   children,
   ...props
 }: ChakraLinkAndNextProps) => {
-  const linkProps = {
-    prefetch:
-      String(href).includes('create-wiki') || String(href).includes('about')
-        ? false
-        : prefetch,
-  }
   return (
-    <NextLink href={href} passHref {...{ linkProps }}>
+    <LinkWrapper href={href} prefetch={prefetch}>
       <ChakraLink {...props}>{children}</ChakraLink>
-    </NextLink>
+    </LinkWrapper>
   )
 }
