+++ src/components/Elements/SocialIcon/SocialIcon.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Button, chakra } from '@chakra-ui/react'
-import Link from '@/components/Elements/Link/Link'
+import Link from '@/components/Elements/LinkElements/Link'
 
 const SocialIcon = ({ Icon, link }: { Icon: JSX.Element; link: string }) => (
   <Link target="_blank" href={link} _hover={{ textDecoration: 'none' }}>
