+++ src/components/Elements/index.ts
@@ -1,9 +1,9 @@
 export { default as Avatar } from './Avatar/Avatar'
-export { default as LinkButton } from './Button/LinkButton'
+export { default as LinkButton } from './LinkElements/LinkButton'
 export { default as Carousel } from './Carousel/Carousel'
 export { default as Dropzone } from './Dropzone/Dropzone'
 export { default as ImageInput } from './ImageInput/ImageInput'
-export { default as Link } from './Link/Link'
+export { default as Link } from './LinkElements/Link'
 export { default as Logo } from './Logo/Logo'
 export { default as Modal } from './Modal/Modal'
 export { default as SocialIcon } from './SocialIcon/SocialIcon'
