+++ src/components/Landing/HeroCard.tsx
@@ -9,7 +9,7 @@ import { WIKI_IMAGE_ASPECT_RATIO } from '@/data/Constants'
 import { WikiImage } from '../WikiImage'
 import DisplayAvatar from '../Elements/Avatar/Avatar'
 import { Link } from '../Elements'
-import LinkOverlay from '../Elements/LinkOverlay/LinkOverlay'
+import LinkOverlay from '../Elements/LinkElements/LinkOverlay'
 
 const CARD_DETAILS_LENGTH = 100
 const HERO_WIKI_IMG_WIDTH = 290
