+++ src/components/Landing/TrendingWikis.tsx
@@ -19,7 +19,7 @@ import { getUsername } from '@/utils/getUsername'
 import { getReadableDate } from '@/utils/getFormattedDate'
 import { useENSData } from '@/hooks/useENSData'
 import { Carousel, Link } from '../Elements'
-import LinkOverlay from '../Elements/LinkOverlay/LinkOverlay'
+import LinkOverlay from '../Elements/LinkElements/LinkOverlay'
 import DisplayAvatar from '../Elements/Avatar/Avatar'
 
 const TRENDING_WIKI_IMG_WIDTH = 300
