+++ src/components/Wiki/WikiPage/InsightComponents/RelatedWikis.tsx
@@ -7,7 +7,7 @@ import { store } from '@/store/store'
 import { WikiImage } from '@/components/WikiImage'
 import { getWikiImageUrl } from '@/utils/getWikiImageUrl'
 import { getWikiSummary, WikiSummarySize } from '@/utils/getWikiSummary'
-import LinkOverlay from '@/components/Elements/LinkOverlay/LinkOverlay'
+import LinkOverlay from '@/components/Elements/LinkElements/LinkOverlay'
 
 export const RelatedWikiCard = ({ wiki }: { wiki: WikiPreview }) => {
   const { id, title } = wiki
