+++ src/components/Wiki/WikiPage/InsightComponents/WikiDetails.tsx
@@ -16,7 +16,7 @@ import { SiIpfs } from 'react-icons/si'
 import { GoLink } from 'react-icons/go'
 import { WikiImage } from '@/components/WikiImage'
 import { Author, BaseCategory, WikiPreview } from '@/types/Wiki'
-import Link from '@/components/Elements/Link/Link'
+import Link from '@/components/Elements/LinkElements/Link'
 import DisplayAvatar from '@/components/Elements/Avatar/Avatar'
 import { useENSData } from '@/hooks/useENSData'
 import config from '@/config'
