+++ src/pages/tags/[tag].tsx
@@ -16,7 +16,7 @@ import { getRunningOperationPromises } from '@/services/categories'
 import { store } from '@/store/store'
 import WikiPreviewCard from '@/components/Wiki/WikiPreviewCard/WikiPreviewCard'
 import { getTagWikis } from '@/services/wikis'
-import Link from '@/components/Elements/Link/Link'
+import Link from '@/components/Elements/LinkElements/Link'
 import { Wiki } from '@/types/Wiki'
 import { useRouter } from 'next/router'
 import { ITEM_PER_PAGE } from '@/data/Constants'
